Pulled up formatInfo method to MutableContainerTreeNode
